requirements.txt
tflite-runtime
Pillow
numpy
tk
